Websocket: Break out all subscription methods into subscriptions.go
